requirements.txt
alembic==1.15.1
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
cssmin==0.2.0
Flask==3.1.0
Flask-Assets==2.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsmin==3.0.1
Mako==1.3.9
MarkupSafe==3.0.2
pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
SQLAlchemy==2.0.39
typing_extensions==4.12.2
uWSGI==2.0.28
webassets==2.0
Werkzeug==3.1.3
WTForms==3.2.1
psycopg2-binary>=2.9.3
python-dotenv>=0.19.0
